requirements.txt
streamlit
agno
google-genai
google-cloud-bigquery
pandas
pandas-gbq
sqlalchemy